requirements.txt
psutil==5.9.5
pyperclip==1.8.2
requests==2.26.0
scapy>=2.5.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
python-dotenv>=1.0.0
adbutils>=1.2.0
loguru>=0.7.0
tkinter>=8.6 